Working with full front-end re-write
