--- clean_code/.gitignore
@@ -1,13 +1,0 @@
-grobid-*
-temp
-.devcontainer
-__pycache__
-.env
-service_account_key.json
-logs
-.vscode
-*.zip
-.devcontainer
-venv
-scratch.*
-cache
