--- clean_code/.streamlit/config.toml
@@ -1,2 +1,0 @@
-[server]
-maxUploadSize = 10
