--- clean_code/README.md
@@ -1,4 +1,0 @@
-# master-project
-My master project repository. I will upload things related to my master project.
-
-GROBID SERVER is running on Huggingface. Please do not misuse it.
