--- clean_code/config.json
@@ -1,6 +1,0 @@
-{
-    "grobid_server": "https://kermitt2-grobid.hf.space",
-    "batch_size": 1000,
-    "sleep_time": 5,
-    "timeout": 60
-}
