--- clean_code/enums.py
@@ -1,11 +1,0 @@
-from enum import auto, StrEnum
-
-class TaskType(StrEnum):
-    DATASET = auto()
-    BASELINE = auto()
-
-
-class SearchEngine(StrEnum):
-    KAGGLE = auto()
-    HF = auto()
-    PwC = auto()
