--- error_codes.py
@@ -1,5 +1,0 @@
-from enum import StrEnum
-
-
-class ERROR_CODES(StrEnum):
-    E01 = "Already loaded!"
