+++ main.py
@@ -2,312 +2,341 @@ from dotenv import load_dotenv
 
 load_dotenv()
 
-from pathlib import Path
-import hashlib
-import shutil
-import utils
-import pipeline as pl
+
 import streamlit as st
-import pydash as py_
-import pandas as pd
-
-# from streamlit_pdf_viewer import pdf_viewer
-from annotate_pdf import annotate_pdf
-from enums import EntityType
-from icecream import ic
-
-# import extra_streamlit_components as stx
-from data_classes import Paper, TaskObject
-from bs4 import BeautifulSoup as bs4
-import time
-from error_codes import ERROR_CODES
-
-from multiprocessing import Pool
-
-
-LLM_TEMPERATURE = 0.06
-
 
 st.set_page_config(
-    page_title="Dataset Mention Extraction",
-    page_icon="🧊",
-    layout="centered",
-    initial_sidebar_state="expanded",
+    page_title="Entity Mention Extraction", page_icon="⛏️", layout="centered"
+)
+
+import time
+from pathlib import Path
+
+import pandas as pd
+import pydash as py_
+from enums import TaskType
+from icecream import ic
+from models import Load_XML, Paper, Task, Upload_PDF
+from streamlit.runtime.uploaded_file_manager import UploadedFile
+from utils import (
+    chcksum,
+    check_for_xmls,
+    download_pdfs,
+    forker,
+    load_xml,
+    pdfs_to_xmls,
+    task_wrapper_extract_entities,
+    upload_convert,
 )
 
 
-if "load_btn" in st.session_state and st.session_state.load_btn == True:
-    st.session_state.disable_load_btn = True
-else:
-    st.session_state.disable_load_btn = False
+def load_pdf_uploads(pdfs: list[UploadedFile]):
+    pdfs_chain = (
+        py_.chain(pdfs)
+        .filter_(lambda pdf: isinstance(pdf, UploadedFile))
+        .map_(lambda pdf: Upload_PDF(chcksum(pdf.getbuffer()), pdf))
+        .uniq_with(lambda file_1, file_2: file_1.id == file_2.id)
+        .filter_(lambda pdf: pdf.id not in st.session_state.papers)
+    )
 
-if "extract_btn" in st.session_state and st.session_state.extract_btn == True:
-    st.session_state.disable_extract_btn = True
-else:
-    st.session_state.disable_extract_btn = False
-
-if "xmls" not in st.session_state:
-    st.session_state.xmls = []
-
-if "warnings" not in st.session_state:
-    st.session_state.warnings = {}
-
-if "tasks" not in st.session_state:
-    st.session_state.tasks = {}
-else:
-    st.session_state.tasks = (
-        py_.chain(st.session_state.tasks.items())
-        .filter_(lambda task: task[1].pending == False)
-        .from_pairs()
+    papers_from_disk = (
+        pdfs_chain.map_(lambda pdf: Load_XML(pdf.id, check_for_xmls(pdf.id)))
+        .map_(load_xml)
         .value()
     )
-    
-if "exec_time" not in st.session_state:
-    st.session_state.exec_time = 0.0
 
+    for paper in papers_from_disk:
+        if paper == None:
+            continue
+        st.session_state.papers[paper.id] = paper
 
-def xml_loader(xml_files):
-    xmls = []
-    for xml_id, xml_path in xml_files.items():
-        title = ""
-        with open(xml_path, "r") as file:
-            title = bs4(file, features="lxml").title
-            if title == None:
-                title = xml_id
-            else:
-                title = py_.human_case(title.get_text())
-        xmls.append(Paper(xml_path, title, xml_id))
-    return xmls
-
-
-
-def task_wrapper_extract_entities(task: TaskObject):
-    start = time.perf_counter()
-    task.extracted_ents = utils.extract_entities(
-        verify=task.verify,
-        chunks=task.chunks,
-        temperature=task.temperature,
-        entity_type=task.entity_type.lower(),  # type: ignore
-        q_embeds=task.query_embeds,  # type: ignore
+    to_upload = pdfs_chain.difference_with(
+        st.session_state.papers.keys(),
+        lambda upload_pdf, paper_id: upload_pdf.id == paper_id,
     )
-    task.time_elapsed = time.perf_counter() - start
-    task.pending = False
-    return task
+    if to_upload.apply(len).value() < 1:
+        st.rerun()
+
+    try:
+        papers_from_upload = upload_convert(to_upload).map_(load_xml).value()
+        for paper in papers_from_upload:
+            if paper == None:
+                continue
+            st.session_state.papers[paper.id] = paper
+    except Exception as e:
+        st.session_state.errors.append(e)
+        st.rerun()
 
 
-def forker(tasks: list[TaskObject], function):
-    start = time.perf_counter()
-    
-    with Pool() as pool:
-        results = pool.imap_unordered(function, tasks)
-        
-        for result in results:
-            st.session_state.tasks[result.task_id] = result
-    
-    st.session_state.exec_time = time.perf_counter() - start
+def load_pdf_downloads(urls: list[str]):
+    ic(urls)
+    downloaded_pdfs = download_pdfs(urls).value()
+    pdf_cache_dir = None
+    for pdf in downloaded_pdfs:
+        if pdf.id in st.session_state.papers:
+            pdf.file_path.unlink()
+        elif (xml_path := check_for_xmls(pdf.id)) != None:
+            pdf.file_path.unlink()
+            # papers_from_disk
+            st.session_state.papers[pdf.id] = load_xml(
+                Load_XML(id=pdf.id, path=xml_path)
+            )
+        else:
+            pdf_cache_dir = pdf.file_path.parent
 
+    if pdf_cache_dir == None or len(list(pdf_cache_dir.glob("*.pdf"))) < 1:
+        st.rerun()
 
-
-def chcksum(buffer):
-    if isinstance(buffer, str):
-        buffer = buffer.encode("utf-8")
-    return hashlib.sha256(buffer).hexdigest()
+    try:
+        papers_from_download = pdfs_to_xmls(pdf_cache_dir).map_(load_xml).value()
+        for paper in papers_from_download:
+            if paper == None:
+                continue
+            st.session_state.papers[paper.id] = paper
+    except Exception as e:
+        st.session_state.errors.append(e)
+        st.rerun()
 
 
 def main():
-    files = st.file_uploader(
-        "Upload research papers", accept_multiple_files=True, type="pdf"
-    )
+    upload_method, download_method = st.tabs(["Upload", "URL"])
 
-    st.button("Load PDFs", key="load_btn", disabled=st.session_state.disable_load_btn)
+    with upload_method:
+        pdfs = st.file_uploader(
+            "Upload research papers",
+            accept_multiple_files=True,
+            type="pdf",
+            disabled=st.session_state.disable_load_btn,
+        )
+        if pdfs != None and len(pdfs) > 0:
+            st.button(
+                f"Upload PDF{'s' if len(pdfs) > 1 else ''}",
+                key="upload_btn",
+                disabled=st.session_state.disable_load_btn,
+            )
+
+    with download_method:
+        edited_df = st.data_editor(
+            pd.DataFrame([{"url": None}]),
+            column_config={
+                "url": st.column_config.LinkColumn(
+                    label="URL",
+                    width="large",
+                    validate=r"^https:\/\/.+$",
+                    display_text=r"^https:\/\/.+?\/([^\/]+?)$",
+                )
+            },
+            use_container_width=True,
+            disabled=st.session_state.disable_load_btn,
+        ).dropna()
+        if len(edited_df) > 0:
+            st.button(
+                f"Download PDF{'s' if len(edited_df) > 1 else ''}",
+                key="download_btn",
+                disabled=st.session_state.disable_load_btn,
+            )
 
     if st.session_state.disable_load_btn:
-        files = (
-            py_.chain(files or [])
-            .map_(lambda file: (chcksum(file.getbuffer()), file))
-            .uniq_with(lambda file_1, file_2: file_1[0] == file_2[0])
-            .map_(lambda file: file[1])
-        )
-
-        XML_CACHE = [
-            xml_p.name.replace(".grobid.tei.xml", "")
-            for xml_p in Path("temp/xmls").glob("**/*.grobid.tei.xml")
-        ]
-
-        temp_xmls = files.filter_(
-            lambda pdf: chcksum(pdf.getbuffer()) not in XML_CACHE
-        ).value()
-
-        persisted_xmls = py_.pick(
-            {
-                xml_p.name.replace(".grobid.tei.xml", ""): xml_p
-                for xml_p in Path("temp/xmls").glob("**/*.grobid.tei.xml")
-            },
-            *files.difference(temp_xmls)
-            .map_(lambda file: chcksum(file.getbuffer()))
-            .value(),
-        )
-
-        if len(temp_xmls) < 1:
-            st.session_state.warnings[ERROR_CODES.E01] = "PDFs are already loaded!"
-            for xml in xml_loader(persisted_xmls):
-                st.session_state.xmls.append(xml)
-            st.rerun()
-
-        new_pdf_dir = utils.upload_pdfs(temp_xmls)
-        if new_pdf_dir == None:
-            st.rerun()
-        xml_files = utils.pdfs_to_xmls(new_pdf_dir)
-        if xml_files == None:
-            st.rerun()
-
-        for xml in xml_loader(xml_files):
-            st.session_state.xmls.append(xml)
-
+        if pdfs != None and len(pdfs) > 0 and st.session_state.upload_btn:
+            ic("upload")
+            load_pdf_uploads(pdfs)
+        if len(edited_df["url"]) > 0 and st.session_state.download_btn:
+            ic("download")
+            load_pdf_downloads(edited_df["url"].to_list())
         st.rerun()
 
-    if len(st.session_state.xmls) > 0 and not st.session_state.disable_load_btn:
-        if ERROR_CODES.E01 in st.session_state.warnings:
-            st.warning(st.session_state.warnings[ERROR_CODES.E01])
-            del st.session_state.warnings[ERROR_CODES.E01]
-        else:
-            st.success("Succesfully loaded!", icon="✅")
+    if len(st.session_state.errors) > 0:
+        for err in st.session_state.errors:
+            st.error(err)
+        st.session_state.errors = []
 
-    for xml in st.session_state.xmls:
-        expander = st.expander(py_.human_case(xml.title))
-        expander.write(xml.paper_id)
-        
-        include = expander.checkbox("Include", value=True, key=f"chkbox-{xml.paper_id}")
+    if st.session_state.exec_time != None:
+        st.success(f"Total time taken: {st.session_state.exec_time}s")
 
-        if not include:
-            continue
+    for paper in st.session_state.papers.values():
+        paper: Paper = paper
+        expander = st.expander(py_.human_case(paper.title))
 
-        tasks = expander.multiselect(
-            "Entities to extract",
-            list(EntityType),
-            EntityType.DATASET,
-            py_.human_case,
-            f"mltislct-{xml.paper_id}",
-            "More extractable entities can be added",
-        )
+        task_tab, result_tab = expander.tabs(["Tasks", "Results"])
 
-        task_ids = {task: chcksum(f"{xml.paper_id}-{task.lower()}") for task in tasks}
+        with task_tab:
 
-        status_codes = ["🟡", "🔴", "🟢"]
+            include = task_tab.checkbox("Include", value=True, key=f"chkbox-{paper.id}")
 
-        status = {}
-        att = {}
+            if not include:
+                exit(task_tab)
 
-        for task in tasks:
-            if (
-                task_ids[task] not in st.session_state.tasks
-                or st.session_state.tasks[task_ids[task]].pending
-            ):
-                status[task] = status_codes[0]
-                att[task] = None
-            elif (
-                task_ids[task] in st.session_state.tasks
-                and st.session_state.tasks[task_ids[task]].pending == False
-            ):
-                status[task] = status_codes[2]
-                att[task] = st.session_state.tasks[task_ids[task]].time_elapsed
+            task_types = task_tab.multiselect(
+                "Entities to extract",
+                list(TaskType),
+                TaskType.DATASET,
+                py_.human_case,
+                f"mltislct-{paper.id}",
+                "More extractable entities can be added",
+            )
 
-        task_df = expander.data_editor(
-            pd.DataFrame(
+            potential_task_ids = {
+                task_type: chcksum(f"{paper.id}-{task_type}")
+                for task_type in task_types
+            }
+
+            status_colors = ["🟡", "🔴", "🟢"]
+
+            status = {}
+            att = {}
+
+            for task_id in potential_task_ids.values():
+                if (
+                    task_id in st.session_state.tasks
+                    and not (task := st.session_state.tasks[task_id]).pending
+                ):
+                    status[task_id] = status_colors[2]
+                    att[task_id] = task.time_elapsed
+                    continue
+
+                status[task_id] = status_colors[0]
+                att[task_id] = None
+
+            task_df = pd.DataFrame(
                 [
                     {
-                        "task": py_.human_case(task),
+                        "task_type": py_.human_case(task_type),
                         "verify": False,
-                        "status": status[task],
-                        "att": att[task],
+                        "status": status[task_id],
+                        "att": att[task_id],
                     }
-                    for task in tasks
+                    for task_type, task_id in potential_task_ids.items()
                 ]
-            ),
-            column_config={
-                "task": st.column_config.TextColumn(
-                    label="Extractable Entities", disabled=True, width="large"
-                ),
-                "verify": st.column_config.CheckboxColumn(
-                    label="Verify Entities", default=True
-                ),
-                "status": st.column_config.TextColumn(label="Status", disabled=True),
-                "att": st.column_config.TextColumn(label="ATT", disabled=True),
-            },
-            hide_index=True,
-            key=f"data-ed-{xml.paper_id}",
-            use_container_width=True,
-        )
-
-        text = utils.xml_to_body_text(xml.path)
-        chunks = pl.chunker(text)
-                
-
-        for row in task_df.T:
-            if task_df["status"][row] == "🟢":  # type: ignore
-                continue
-            st.session_state.tasks[
-                chcksum(f"{xml.paper_id}-{task_df['task'][row].lower()}")  # type: ignore
-            ] = TaskObject(
-                task_id=task_ids[task_df["task"][row].lower()],  # type: ignore
-                paper_id=xml.paper_id,
-                entity_type=task_df["task"][row].lower(),  # type: ignore
-                verify=task_df["verify"][row],  # type: ignore
-                xml_path=xml.path,
-                chunks=chunks,
-                query_embeds=utils.query_embedder(task_df["task"][row].lower()),  # type: ignore
-                temperature=LLM_TEMPERATURE,
             )
-        
-        
 
-    if len(st.session_state.tasks) > 0:
+            edited_task_df = task_tab.data_editor(
+                task_df,
+                column_config={
+                    "task_type": st.column_config.TextColumn(
+                        label="Extractable Entities", disabled=True, width="large"
+                    ),
+                    "verify": st.column_config.CheckboxColumn(
+                        label="Verify Entities", default=True
+                    ),
+                    "status": st.column_config.TextColumn(
+                        label="Status", disabled=True
+                    ),
+                    "att": st.column_config.TextColumn(
+                        label="ATT", disabled=True, help="Actual Time Taken"
+                    ),
+                },
+                hide_index=True,
+                key=f"data-ed-{paper.id}",
+                use_container_width=True,
+            )
+
+            for row in range(len(edited_task_df)):
+                ID = potential_task_ids[edited_task_df["task_type"][row].lower()]
+
+                if (
+                    ID in st.session_state.tasks
+                    and not st.session_state.tasks[ID].pending
+                ):
+                    continue
+
+                st.session_state.tasks[ID] = Task(
+                    paper=paper,
+                    id=ID,
+                    task_type=edited_task_df["task_type"][row].lower(),
+                    verify=edited_task_df["verify"][row],
+                )
+
+        with result_tab:
+            st.write()
+
+    if (
+        py_.chain(st.session_state.tasks.values())
+        .filter_(lambda task: task.pending)
+        .apply(len)
+        .value()
+        > 0
+    ):
+        ic(
+            py_.chain(st.session_state.tasks.values())
+            .filter_(lambda task: task.pending)
+            .value()
+        )
         st.button(
             "Extract Entities",
             key="extract_btn",
-            disabled=st.session_state.disable_extract_btn
-            or len(
-                py_.chain(st.session_state.tasks.values())
-                .filter_(lambda task: task.pending)
-                .value()
-            )
-            < 1,
+            disabled=st.session_state.disable_extract_btn,
         )
 
     if st.session_state.disable_extract_btn:
-        
-        forker(
-            py_.chain(st.session_state.tasks.values())
-            .filter_(lambda task: task.pending)
-            .value(),
-            task_wrapper_extract_entities,
+        tasks: list[Task] = list(st.session_state.tasks.values())
+        updated_tasks, exec_time = forker(
+            [task for task in tasks if task.pending], task_wrapper_extract_entities
         )
 
+        st.session_state.tasks = {
+            **st.session_state.tasks,
+            **{ut.id: ut for ut in updated_tasks},
+        }
+
+        st.session_state.exec_time = exec_time
         st.rerun()
 
-    if (
-        not st.session_state.disable_extract_btn
-        and len(
-            py_.chain(st.session_state.tasks.items())
-            .filter_(lambda task: task[1].pending)
-            .value()
-        )
-        != 0
-    ):
-        st.toast("Extracted Successfully", icon="✅")
-
-    st.write(f"Took {st.session_state.exec_time}s.")
 
 if __name__ == "__main__":
+    ic.configureOutput(prefix=lambda: "%i |> " % int(time.time()))
+
     Path("temp/pdfs/").mkdir(exist_ok=True, parents=True)
     Path("temp/xmls/").mkdir(exist_ok=True, parents=True)
-    for entity_type in EntityType:
+    for entity_type in TaskType:
         Path(f"temp/views/{entity_type}").mkdir(exist_ok=True, parents=True)
     Path("cache/transformers/model/").mkdir(exist_ok=True, parents=True)
-    
-    ic.configureOutput(prefix=lambda: '%i |> ' % int(time.time()))
+
+    if (
+        "upload_btn" in st.session_state and st.session_state["upload_btn"] == True
+    ) or (
+        "download_btn" in st.session_state and st.session_state["download_btn"] == True
+    ):
+        st.session_state["disable_load_btn"] = True
+    else:
+        st.session_state["disable_load_btn"] = False
+
+    if "extract_btn" in st.session_state and st.session_state["extract_btn"] == True:
+        st.session_state["disable_extract_btn"] = True
+    else:
+        st.session_state["disable_extract_btn"] = False
+
+    if "papers" not in st.session_state:
+        st.session_state.papers = {}
+
+    if "errors" not in st.session_state:
+        st.session_state.errors = []
+
+    if "tasks" not in st.session_state:
+        ic("here")
+        st.session_state.tasks = {}
+    else:
+        st.session_state.tasks = (
+            py_.chain(st.session_state.tasks.items())
+            .filter_(lambda task: task[1].pending == False)
+            .from_pairs()
+            .value()
+        )
+
+    if "exec_time" not in st.session_state:
+        st.session_state.exec_time = None
 
     main()
 
-    # del st.session_state.tasks
-    # shutil.rmtree("temp")
+
+# A file can exists:
+# ✅ in memory (upload case)
+#    - use py_.uniq to get uniq entires
+# ✅ in session_state (uploaded): st.session_state.upload_pdfs
+#    - filter pdfs already in st.session_state.upload_pdfs
+# ✅ on disk (uploaded and saved; downloaded)
+#    - already processed (constraint)
+#       + if not processed then remove the cached pdf and save the new pdf
+#    - pdfs already on disk
+#    - not the job here
+#    - handle in upload_pdf function
+# ✅ save the pdf and load in disk
