+++ texts.py
